Switch default dissimilarity from Jaccard to Dice
The number of overlapping feature sets is non-linear for a threshold tau
on the Jaccard distance. E.g., for 10 features and tau in 0.1, 0.2, ..., 1,
no tau creates an overlap of exactly 2 (tau=0.8 allows overlap of 3,
tau=0.9 allows overlap of 1), while tau=0.2 yields the same result as
tau=0.3 (overlap of 8). This might make comparing and interpreting results
over tau difficult. Thus, we switch to the Dice coefficient, which is
linear in the set overlap. We leave Jaccard as a non-default option.