Adapt experimental design and parallelization
Since the refactoring in 301236ca4b, runtime might benefit from
initializing a feature selector once and then invoking it multiple times.
Thus, we change the pipeline from parallelizing over

- datasets (while all CV folds of a dataset are considered sequentially)
- feature-selection methods
- search configurations for alternatives

to paralellizing over

- datasets (as before)
- CV folds (new, might help to load-balance large datasets)
- feature-selection methods (as before)

In particular, for one feature-selection method on a data split, run all
configurations of alternatives-search sequentially.

Besides this performance-oriented refactoring, we also make changes to the
experimental design, based on insights from a preliminary run of the
experimental pipeline:

- evaluate less alternatives for simultaneous search than for sequential
search (as the former scales horribly with the number of alternatives)
- exclude the greedy wrapper, which is very slow (due to many solver calls)
- adapt values for threshold "tau" to "k" (i.e., if "k" larger, then it
makes sense to evaluate more values for "tau")