Introduce optimization timeout
A few optimization configurations take disproportionally long compared to
the other ones. To still allow for a broad experimental design (and keeping
it roughly a cross-product of experimental dimensions rather than cutting
down the latter manually), we introduce a timeout on the optimizer runs.
Search for alternatives might still take longer if the optimizer is invoked
multiple times, as it is in the sequential search procedure and for greedy
wrapper feature selection; however, both of these methods have a hard limit
on their iteration count, so the overall runtime is still under control).