Fix indexing in wrapper's main loop
Inspired by the pseudo-code, we started looping from index 1 instead of 0
for swapping features in the wrapper feature-selection procedure.
Though that does not impact functionality strongly (apart from the
consequence that the first feature was never forced to be swapped, though
its selection could still be changed when swapping another feature),we fix
it now.