Allow absolute values for tau
For set dissmilarity, we currently use the Dice coefficient. Also, both
sets have the same size k. Thus, one can easily translate the distance tau
from [0, 1] to an absolute number of differing features, i.e., tau * k.
Since e6ce85bf48, we already chose values of tau seemingly corresponding
to integer numbers of differing features. However, thanks to floating point
arithmetics, this did not always hold:

```
[(1- tau) * 5 for tau in [0.2, 0.4, 0.6, 0.8, 1]]
```

Output:

```
[4.0, 3.0, 2.0, 0.9999999999999998, 0]
```

Even slight deviations might be problematic, because an overlap of 0.9999
features still is the same an an overlap of 0 features (unless the solver
messes up the domain of the binary selection variables and allows them to
assume values slightly less than one).
Thus, we adapt the dissimilarity method to also accept absolute values for
tau, and we only pass such integer values in the experimental pipeline.
For one example we tested locally, this also prevented a mysterious crash
of the solver (we found the floating-point bug investigation that solver
crash). Hopefully that helps for the rest of the pipeline as well.