Try Gekko for optimization
Allows to formulate product terms without interaction variables, which is a
plus for simultaneous search with many alternatives. However, seems to have
more overhead than "mip", i.e., is slower for easy optimization tasks (of
which we have quite a lot). Is a lot slower for sequential search. Also,
the train objectives in the results for sequential search are not
monotonically decreasing (even if we already tried adapting some numerical
tolerances). Also, getting the value from the expressions for the
objectives did not work properly (0 returned), so we needed to compute the
objectives manually.