Try CVXPY for optimization
which uses the GLPK_MI optimizer (others are possible as well). Compared
to "mip", problems are immutable, so we have to do some copying when
adding further constraints. Also, we need to store some constraints
before creating the problem (we can only add them in bulk, not iteratively).
So code-wise, this is not an improvement. Runtimes seemed comparable in a
small experiment locally, but the pipeline barely made any progress running
half a day on the server.