Try mRMR for feature selection
Another multivariate filter method (we already have FCBF). Big disadvantage
that it needs interaction variables for feature-feature relationships, so
it scales badly (like simultaneous search - combining these two would be
an overkill for the pipeline). For example, running sequential search on a
dataset with 16 features took 13 s (which is slow), while 30 features even
required 3 min (and we have datasets a lot bigger than that). In contrast,
FCBF models feature interactions with constraints instead of pairwise
interaction terms, so it is much faster (though slower than univariate
selection), but there often is no solution satisfying the constraints.