Try Google OR-Tools for optimization
OR-Tools has two modules suitable for our problem: one for CSPs and one
for LPs. However, the CSP-one only supports one solver, which can also
be used for LPs, and does not allow setting a solver timeout. Thus, we use
the LP module. This supports various free (including our friend CBC) and
commercial solvers. Syntax is not that different to PuLP/python-mip
(though sometimes a bit strange and not well-documented). A disadvantage is
that removing constraints is not allowed, but we can fix variables to
values (which suffices for our use case in the greedy wrapper) or change
coefficients in existing constraints.