Add mRMR feature selector
We already tried that in the "experimental" branch a while ago (11acaa66),
though with Python-MIP instead of OR-Tools. Still, the code added now is
quite similar to the old one. However, we found a paper that presents a
more efficient encoding of interaction variables for mRMR and other
multivariate filter feature selectors, so we should have better runtimes on
large datasets now.

Besides adding the new feature selector as a separate class, we also need
to do some refactoring:
- pass the number of features "k" already during solver initialization
(rather than during search), as the objective function of mRMR (and
potentially other feature selectors as well) depends on it
- add a new abstract class for white-box feature selection (which is the
superclass of the new mRMR selector and the old abstract class for
univariate selectors ("LinearQualityFeatureSelector"))
- unrelated: minor refacotring in GreedyWrapperSelector based on pylint