Add num-alternative plots with infeasible objective set to 0
Plots showing the trend of the objective over the number of alternatives
are central to our work. Before, we only showed aggregate (box or median)
quality that automatically excluded infeasible feature sets (as their
quality is NaN). However, it also makes sense to consider the objective of
infeasible feature sets to be 0 (i.e., like no feature has been selected,
at least for the MI selector we use for plotting). Thus, we add such plots
to complement the existing plots. This triggered some refactoring and
further harmonization, like
- write normalized evaluation metrics to separate data frames instead of
  appending them to the overall results
- try to copy as few parts of the overall data frame as possible
- always consider "k=10"-results in the num-alternatives (+/- tau) analysis
- change the order of some console outputs
- show mean instead of median objective in num-alternatives-tau plots
  (median looks a bit degenerate in the "fillna" version of the plot, as
  over half the feature sets are infeasible and thus have objective of 0)