Replace Cbc with SCIP
We changed the solver package from "mip" to "ortools" in commit d7fa31da2d.
This introduced a newer version of the same solver ("Cbc") and removed the
few buggy results (wrong number of features selected) that occurred with
"mip". However, even with the new package, the solver silently crashed in a
few of the parallel tasks. By silent, I mean that you could only see that
these tasks didn't finish while comparable tasks finished ages ago, but
there was no error message in the parallelized pipeline). The crashed
re-occurred when running the same tasks again. However, in some cases, the
pipeline crashed only on the server but not locally (maybe because
different operating systems and potentially different solver binaries). For
the current pipeline, the following (3 out of 1500) tasks were affected:

- dataset: "churn", fold: 3, FS: MRMR
- dataset: "coil2000", fold: 2, FS: MRMR
- dataset: "GAMETES_Epistasis_2_Way_20atts_0.1H_EDM_1_1", fold: 3, FS: MRMR

In the previous experimental pipeline (commit d9d2f6c966, run on May 23),
a different (1 out of 1200) task was affected:

- dataset: "spambase", fold: 0, FS: ModelImportance

While the latter task could still be run locally, that would be too
expensive for the mRMR tasks (in particular, as we also increased the
solver timeout for simultaneous search and added another search setting).
Also, running tasks on different machines messes up the solver timeouts and
runtime results.

Fortunately, Google OR-Tools allows exchanging the solver easily. Thus, we
replace Cbc with SCIP. The latter runs all 1500 tasks from the current
pipeline without crashes and is also faster (141 h vs. 160.5 h on the
chair's 32-core server).