Clear swap-variables list in Greedy Wrapper
When we replaced Python-MIP with OR-Tools a long time ago (d7fa31da2d), we
introduced a minor bug in the Greedy Wrapper: We never cleared the list of
swap variables, so it continously grew over iterations. However, the list
does not determine which variables are actually swapped, but only which
variables are reset to their default range after swapping. For variables
currently not swapped, this is an operation without any effect (at least
as far as I know), as it sets their new range to their old range. We only
noticed this bug now when debugging the Greedy Wrapper, as the latter
produced significantly different results when using SCIP instead of Cbc as
solver (which probably is not related to the bug we fix here).