Fix signature of sequential search
The refactoring in 104653f5b1 removed an unused argument from sequential
search, i.e., the objective aggregation, which only matters for
simultaneous search. However, the call of the search routine in the
experimental pipeline, i.e., in `run_experiments.py#evaluate_one_search()`,
always passes a value for the objective aggregation, since the call does
not discriminate between sequential and simultaneous search. Thus, we
re-introduce this argument in sequential search to prevent the pipeline
from crashing.