Adapt evaluation to non-occurring optimization statuses
While the previous pipeline results (from December 2022) contained a few
instances of the optimization status "Not solved" for mRMR, the recent
results only contain this status for Greedy Wrapper. Thus, we need to adapt
the code producing frequency tables of optimization statuses for all
white-box feature-selection methods, because we previously assumed that all
these statuses will occur (and therefore form a column in the table).