Add class for manually defined feature qualities
Not relevant for our experiments. However, if one wants solve univariate-
quality examples without providing a dataset (as we did in the section on
the objective-aggregation operators for simultaneous search), this is the
way to go. We tried the two examples from our paper and found that our
manual computation of the objective values is correct :-)