Add greedy-balacing search
While the greedy-replacement search introduced in f1f21ddcdd employs a
sequential search procedure, the new search is simultaneous. Like greedy
replacement, it also is intended for univariate qualities and retains the
same top features in each set, but uses a heuristic for "balanced number
partitioning" for the remaining features. For consistency reasons, we also
refactor the existing greedy-replacement search slightly.