Evaluate correlation of normalized quality with "n"
Requested by a reviewer. The arXiv evaluation already has a correlation
analysis of "n" and feature-set quality, but this analysis does not
consider the parameters "a" and "tau" from the search for alternatives.
Thus, we extend the user-parameter evaluation loop (that also creates the
plots over "a" and "tau") to output some correlations to the console.