Try kNN as prediction model
Requested by a journal reviewer. By default, predictions are directly made
after the search for alternatives. Thus, if we appended the list of
predictions models in "predictions.py", we would need to run the full
experimental pipeline (10-11 days). For a preliminary evaluation, we
instead load the list of found feature sets from the existing experimental
results and directly evaluate kNN with these feature sets.
The resulting experimental pipeline is a modified copy of the main
experimental pipeline, cutting away the search for alternatives and
evaluating one (hard-coded) prediction model on existing results,
overwriting the corresponding file by adding two further columns (train and
test MCC of kNN).
We also adapt the evaluation accordingly.