Add search heuristics to experimental pipeline
... though they only work for univariate filter FS, i.e., "MI" and "Model
Gain" in the paper. "Greedy Replacement" is a sequential heuristic,
"Greedy Balancing" a simultaneous one; both are evaluated with the same
parameters as the corresponding exact search methods for alternatives.
The results of the heuristics are appended to the results of the exact
search, since pipeline tasks are only defined by dataset, cross-validation
fold, and feature-selection method, i.e., running all search configurations
for alternatives sequentially instead of producing separate results files.
This means we need to re-run the pipeline for "MI" and "Model Gain" to
get results for the heuristics.