Only use solver-based search for some arXiv evaluations
When comparing the impact of search methods, "tau", or "a", the evaluation
implicitly selects solver-based search methods by subsetting the DataFrame
"results" accordingly. However, other parts of the evaluation (particularly
for the Appendix) directly worked on "results" and were therefore impacted
by the introduction of heuristic search though they shouldn't. In
particular, heuristic search was only evaluated for some feature-selection
methods and should not be included in comparisons of all feature-selection
methods. As a remedy, we introduce a new variable for solver-based search
results and use it where appropriate.