Harmonize plot colors for diss
Within AFS and also relative to other chapters/papers. We already relied
on mostly on sequential palette, which we replace with another sequential
palette now; such palettes are good for grayscale printing. However, we
keep the diverging palette for the metrics-correlation plot and we switch
to a quantitative palette for two multi-category lineplots (on screen,
lines easier to distinguish if diverse colors used; not an issue for
boxplots and for the lineplots where the hue has a numeric interpretation
anyway).