Remove dataset evaluation from diss evaluation
Only relevant insight is that feature-set quality differs between datasets,
which is not specific to alternative feature selection. This evaluation
type is not part of the journal evaluation either.
Also, add section numbers to all printed headings for the experimental
design (for the evaluation, we'll add these numbers whenever we refactor
a section).