Make Greedy Wrapper optimize solution similarity for repair
Until now, the feature-selection method "Greedy Wrapper" used a solver to
turn a potentially invalid solution (in particular, the previous best
solution + two enforced feature swaps) into a valid one. However, this
approach did not explicitly guarantee the new solution to be in any way
similar to the previous solution. At most, we could hope that the solver
chooses a similar solution since it warm-starts from the previous one.
We now explicitly foster similarity by optimizing the Hamming similarity
of the new solution (binary feature-selection vector) to the previous one.